test `traverse` and `status`
